// apps/client/components/types.ts
enum NavigationButtonType {
  Home = "Home",
  Settings = "Settings",
  Money = "Money",
  Message = "Message",
  Contacts = "Contacts",
  Clock = "Clock",
  Shop = "Shop",
  Cart = "Cart",
  Menu = "Menu",
  Search = "Search",
  Tool = "Tool",
  UserAdd = "UserAdd",
  Plus = "Plus",
  Close = "Close",
  ListStar = "ListStar",
  Bag = "Bag",
  CreditCard = "CreditCard",
  RadioGrid = "RadioGrid",
  Payment = "Payment",
  Radio = "Radio",
  Logout = "Logout",
  ArrowBack = "ArrowBack",
  Table = "Table",
  Fork = "Fork",
  Dashboard = "Dashboard"
}

export { NavigationButtonType };